feat: renamed bonnesboites -> lbacompany.service
