+++ server/src/http/controllers/jobs/jobs.controller.ts
@@ -27,7 +27,7 @@ import { delay } from "../../../common/utils/asyncUtils.js"
 import { ICredential } from "../../../common/model/schema/credentials/credential.types.js"
 import { IApiError } from "../../../common/utils/errorManager.js"
 import { ILbaItem } from "../../../services/lbaitem.shared.service.types.js"
-import { getCompanyFromSiret } from "../../../services/bonnesBoites.js"
+import { getCompanyFromSiret } from "../../../services/lbacompany.service.js"
 import { addOffreDetailView, addOffreSearchView, getLbaJobById } from "../../../services/lbajob.service.js"
 import { getPeJobFromId } from "../../../services/pejob.service.js"
 import { getJobsQuery } from "../../../service/poleEmploi/jobsAndCompanies.js"
